+++ packages/core/src/components/ozon-content/nodes/inline-tekst-afbeelding.node.tsx
@@ -7,15 +7,15 @@ export class OzonContentInlineTekstAfbeeldingNode implements OzonContentNode {
   name = ["InlineTekstAfbeelding", "Illustratie"];
 
   render(node: Element, { urlResolver }: OzonContentNodeContext) {
-    const value = node.getAttribute("naam");
+    const naam = node.getAttribute("naam");
     const name: "InlineTekstAfbeelding" | "Illustratie" =
       node.tagName === "InlineTekstAfbeelding" ? "InlineTekstAfbeelding" : "Illustratie";
-    const src = urlResolver ? urlResolver(name, "naam", value, node) : value;
+    const src = urlResolver && naam ? urlResolver(name, "naam", naam, node) : naam;
 
     return (
       <img
         src={src ?? undefined}
-        alt={node.getAttribute("naam") ?? undefined}
+        alt={naam ?? undefined}
         height={node.getAttribute("hoogte") ?? undefined}
         width={node.getAttribute("breedte") ?? undefined}
       />