+++ packages/core/src/components/ozon-content/ozon-content.interfaces.ts
@@ -23,15 +23,10 @@ export interface OzonContentMarkItemHighlightEvent {
 export type OzonContentWijzigActie = "voegtoe" | "verwijder";
 
 export interface OzonContentUrlResolver {
-  (name: "Illustratie" | "InlineTekstAfbeelding", attribute: "naam", value: string | null, element: Element): string;
-  (
-    name: "ExtIoRef" | "ExtRef" | "IntIoRef" | "IntRef",
-    attribute: "ref",
-    value: string | null,
-    element: Element,
-  ): string;
+  (name: "Illustratie" | "InlineTekstAfbeelding", attribute: "naam", value: string, element: Element): string;
+  (name: "ExtIoRef" | "ExtRef" | "IntIoRef" | "IntRef", attribute: "ref", value: string, element: Element): string;
 }
 
 export interface OzonContentBegripResolver {
-  (ref: string | null, element: Element): XMLDocument | string | undefined;
+  (ref: string, element: Element): XMLDocument | string | undefined;
 }