+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## Next
 
+### Changed
+* Ozon Content: `ref` van `OzonContentBegripResolver` en `value` van `OzonContentUrlResolver` heeft `null` type ([#3441](https://github.com/dso-toolkit/dso-toolkit/issues/3441))
+
 ## 🏝️ Release 83.0.0 - 2025-11-27
 
 ### Deprecated